Renamed python_voice_coding_plugin_casterv.py to python_voice_coding_plugin_casterv_v0-5-11 to differentiate Caster version
